# api-docs.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"Very popular Social Network",
      "version":"v1"
   },
   "servers":[
      {
         "url":"http://localhost:50000",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/login":{
         "post":{
            "tags":[
               "auth-controller"
            ],
            "summary":"Authenticate a user",
            "operationId":"loginUser",
            "parameters":[
               {
                  "name":"login",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               },
               {
                  "name":"password",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "400":{
                  "description":"Invalid credentials"
               },
               "200":{
                  "description":"Successful authentication",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/JwtResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/JwtResponse"
                        }
                     }
                  }
               }
            }
         }
      },
      "/register":{
         "post":{
            "tags":[
               "auth-controller"
            ],
            "summary":"Register a user",
            "operationId":"registerUser",
            "requestBody":{
               "description":"Information about user",
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/UserDTO"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful registration",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     }
                  }
               },
               "400":{
                  "description":"Login already exists",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         }
      },
      "/friends":{
         "get":{
            "tags":[
               "friend-controller"
            ],
            "summary":"Get all user's friends",
            "operationId":"getAllFriends",
            "responses":{
               "200":{
                  "description":"All friends returned",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/UsersResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/UsersResponse"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "bearerAuth":[
                     
                  ]
               }
            ]
         }
      },
      "/friends/add":{
         "post":{
            "tags":[
               "friend-controller"
            ],
            "summary":"Add a friend to user's friend list",
            "operationId":"addFriend",
            "parameters":[
               {
                  "name":"userId",
                  "in":"query",
                  "description":"Added user's id",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "400":{
                  "description":"User not found",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               },
               "200":{
                  "description":"Friend has been added to your friends list",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "bearerAuth":[
                     
                  ]
               }
            ]
         }
      },
      "/friends/remove":{
         "delete":{
            "tags":[
               "friend-controller"
            ],
            "summary":"Remove user's friend from friend list",
            "operationId":"removeFriend",
            "parameters":[
               {
                  "name":"userId",
                  "in":"query",
                  "description":"Removed user's id",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Friend has been removed from your friends list",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/Response"
                        }
                     }
                  }
               },
               "400":{
                  "description":"User is not in friends",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "bearerAuth":[
                     
                  ]
               }
            ]
         }
      },
      "/friends/search":{
         "get":{
            "tags":[
               "friend-controller"
            ],
            "summary":"Search people in user's friend list",
            "operationId":"searchFriend",
            "parameters":[
               {
                  "name":"searchQuery",
                  "in":"query",
                  "description":"firstname or lastname of a user",
                  "required":true,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Returned friend list according to query",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/UsersResponse"
                        }
                     },
                     "application/xml":{
                        "schema":{
                           "$ref":"#/components/schemas/UsersResponse"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "bearerAuth":[
                     
                  ]
               }
            ]
         }
      },
...
   },
   "components":{
      "schemas":{
         "JwtResponse":{
            "type":"object",
            "properties":{
               "bearer":{
                  "type":"string"
               }
            }
         },
         "UserDTO":{
            "type":"object",
            "properties":{
               "login":{
                  "type":"string"
               },
               "firstName":{
                  "type":"string"
               },
               "lastName":{
                  "type":"string"
               },
               "password":{
                  "type":"string"
               }
            }
         },
         "Response":{
            "type":"object",
            "properties":{
               "value":{
                  "type":"object"
               }
            }
         },
         "ErrorResponse":{
            "type":"object",
            "properties":{
               "error":{
                  "type":"string"
               }
            }
         },
         "UserResponse":{
            "type":"object",
            "properties":{
               "id":{
                  "type":"integer",
                  "format":"int32"
               },
               "login":{
                  "type":"string"
               },
               "firstName":{
                  "type":"string"
               },
               "lastName":{
                  "type":"string"
               }
            }
         },
         "UsersResponse":{
            "type":"object",
            "properties":{
               "users":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/UserResponse"
                  }
               }
            }
         },
...
      },
      "securitySchemes":{
         "bearerAuth":{
            "type":"http",
            "scheme":"bearer",
            "bearerFormat":"JWT"
         }
      }
   }
}